feat: optional "exact" ignore behaviour
Add a flag to the `ignore` method signature to opt-in to "exact"
path-matching behaviour.

The following vulnerability path: `['foo', 'bar', 'baz']` will be
matched by the following ignore rules under this behaviour:

- 'foo > bar > baz'
- 'foo > *'
- '*'

Under the default matching behaviour, there are some quirks that appear
specific to package manager use cases:

- The first element of vulnerabilities paths is sliced off
- If the path (minus the first element) "contains" the rule as a
  substring, it will match. For example: the rule `file.json` will match
  the path `["dir/file.json"]`.

The Snyk infrastructure-as-code use-case breaks under these default
behaviours, but we still want to use the policy library in order to not
reimplement loading and parsing of the policy file.